+++ example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:example/material_app.dart';
-import 'package:example/cupertino_app.dart';
+// import 'package:example/cupertino_app.dart';
 
 void main() => runApp(MyMaterialApp());
 //void main() => runApp(MyCupertinoApp());